- Moved HeaderMutator with the other header classes
