+++ src/burp/model/header/HeaderMutationGuesser.java
@@ -7,7 +7,6 @@ import burp.IResponseInfo;
 import burp.model.utilities.Utilities;
 import burp.view.ConfigurableSettings;
 import burp.albinowaxUtils.CustomScanIssue;
-import burp.albinowaxUtils.HeaderMutator;
 
 import java.net.URL;
 import java.nio.charset.StandardCharsets;
